+++ NSrtm.Core/NSrtm.Core.csproj
@@ -85,8 +85,10 @@
     <Compile Include="IElevationProvider.cs" />
     <Compile Include="Hgt\DataCells\IHgtDataCell.cs" />
     <Compile Include="Hgt\DataCells\HgtDataCellInvalid.cs" />
+    <Compile Include="Pgm\DataDesciption\PgmDataDescriptionExtractor.cs" />
+    <Compile Include="Pgm\PgmRawDataConverter.cs" />
     <Compile Include="Pgm\PgmGridGraphFactory.cs" />
-    <Compile Include="Pgm\PgmDataDescription.cs" />
+    <Compile Include="Pgm\DataDesciption\PgmDataDescription.cs" />
     <Compile Include="Pgm\GridGraph\PgmGridGraphInFile.cs" />
     <Compile Include="Pgm\GridGraph\PgmGridGraphInMemory.cs" />
     <Compile Include="Pgm\GridGraph\IPgmGridGraph.cs" />
