+++ NSrtm.Core/Pgm/GridGraph/IPgmGridGraph.cs
@@ -1,6 +1,6 @@
 ﻿namespace NSrtm.Core.Pgm.GridGraph
 {
-    public interface IPgmGridGraph
+    internal interface IPgmGridGraph
     {
         double GetUndulation(double latitude, double longitude);
         PgmDataDescription Parameters { get; }
